// src/API/src/db/occurrence/dto/search-occurrence-query.dto.ts
export interface SearchOccurrenceQueryDto {
  id?: string;
  country?: string;
  species?: string;
  season?: string;
  insecticide?: string;
  absence?: string;
  abundance?: string;
  bionomics?: string;
  adult?: string;
  larval?: string;
  timeline?: string;
  take?: string;
  skip?: string;
  fields?: string;
}